requirements.txt
torch==1.11.0+cu113
torchvision ==0.12.0+cu113
onnx==1.13.0
onnxruntime==1.12.1
numpy==1.24.2
scipy==1.8.1

# Solvers
z3-solver==4.8.15.0
gurobipy==10.0.1

# Misc
pandas==1.4.2
Jinja2
tabulate
tqdm==4.63.1
imageio==2.5.0
Pillow==9.1.1
dill==0.3.5.1
